Issue 84: move to 1.6 api
